refactor(subagent): extract SubAgentExecutor interface and add async notification
Split the monolithic delegateSubAgent() method into testable interfaces:
- SubAgentExecutor: encapsulates child runner/session creation, prompt
  building, engine execution, and result extraction behind a mockable
  interface (defaultSubAgentExecutor)
- SubAgentNotifier: queues async completion notifications with a
  thread-safe drain mechanism, injected into the parent turn loop
- Deep-copy WritableRoots/ExecAllowlist/NetworkAllowlist in child runner
  to prevent parent/child config aliasing
