+++ internal/agent/subagent_delegate.go
@@ -15,10 +15,8 @@ import (
 	"time"
 
 	corepkg "github.com/1024XEngineer/bytemind/internal/core"
-	"github.com/1024XEngineer/bytemind/internal/llm"
 	planpkg "github.com/1024XEngineer/bytemind/internal/plan"
 	runtimepkg "github.com/1024XEngineer/bytemind/internal/runtime"
-	"github.com/1024XEngineer/bytemind/internal/session"
 	subagentspkg "github.com/1024XEngineer/bytemind/internal/subagents"
 	"github.com/1024XEngineer/bytemind/internal/tools"
 )
@@ -38,12 +36,6 @@ const (
 	subAgentResultStatusAccepted  = "accepted"
 
 	subAgentRequestedOutputSummary = "summary"
-
-	defaultSubAgentMaxIterations = 8
-
-	subAgentResultPolicyCompressed = `Return your final answer as a single JSON object (no markdown fences). Schema:
-{"summary":"<one-paragraph overview>"}
-Do not include full tool logs.`
 
 	subAgentTaskOutputTool = "task_output"
 	subAgentTaskStopTool   = "task_stop"
@@ -139,6 +131,7 @@ func (r *Runner) delegateSubAgent(
 		metadata["requested_output"] = preflight.RequestedOutput
 	}
 
+	parentSession := execCtx.Session
 	runtimeRequest := RuntimeTaskRequest{
 		SessionID:  sessionIDFromExecutionContext(execCtx),
 		Name:       "delegate_subagent/" + preflightResultName(result.Agent),
@@ -147,7 +140,17 @@ func (r *Runner) delegateSubAgent(
 		Timeout:    preflight.RequestedTimeoutDuration,
 		Metadata:   metadata,
 		Execute: func(taskCtx context.Context) ([]byte, error) {
-			subAgentResult := r.executeSubAgentTask(taskCtx, request, preflight, result.InvocationID, result.Agent, runMode, execCtx)
+			subAgentResult, execErr := r.subAgentExecutor.Execute(taskCtx, SubAgentExecutionInput{
+				Request:      request,
+				Preflight:    preflight,
+				InvocationID: result.InvocationID,
+				Agent:        result.Agent,
+				RunMode:      runMode,
+				ExecCtx:      execCtx,
+			})
+			if execErr != nil {
+				return nil, execErr
+			}
 			output, marshalErr := json.Marshal(subAgentResult)
 			if marshalErr != nil {
 				return nil, &subAgentExecutionError{
@@ -157,6 +160,29 @@ func (r *Runner) delegateSubAgent(
 				}
 			}
 			return output, nil
+		},
+		OnTaskStateChanged: func(task runtimepkg.Task) {
+			if !runtimepkg.IsTerminalTaskStatus(task.Status) || r.subAgentNotifier == nil {
+				return
+			}
+			notification := SubAgentCompletionNotification{
+				ParentSession: parentSession,
+				TaskID:        string(task.ID),
+				Agent:         result.Agent,
+				InvocationID:  result.InvocationID,
+			}
+			if task.Status == corepkg.TaskCompleted {
+				notification.Status = subAgentResultStatusCompleted
+				var parsed tools.DelegateSubAgentResult
+				if jsonErr := json.Unmarshal(task.Output, &parsed); jsonErr == nil {
+					notification.Summary = parsed.Summary
+				}
+			} else {
+				notification.Status = subAgentResultStatusFailed
+				notification.ErrorCode = task.ErrorCode
+				notification.ErrorMessage = fmt.Sprintf("subagent task ended with status %s", task.Status)
+			}
+			r.subAgentNotifier.NotifyCompletion(notification)
 		},
 	}
 
@@ -256,264 +282,6 @@ func (r *Runner) delegateSubAgent(
 		Retryable: true,
 	}
 	return result, nil
-}
-
-func (r *Runner) executeSubAgentTask(
-	ctx context.Context,
-	request tools.DelegateSubAgentRequest,
-	preflight subagentspkg.PreflightResult,
-	invocationID string,
-	agent string,
-	runMode planpkg.AgentMode,
-	execCtx *tools.ExecutionContext,
-) tools.DelegateSubAgentResult {
-	workspace := strings.TrimSpace(r.workspace)
-	if execCtx != nil {
-		if scopedWorkspace := strings.TrimSpace(execCtx.Workspace); scopedWorkspace != "" {
-			workspace = scopedWorkspace
-		}
-	}
-	parentSessionID := ""
-	if execCtx != nil && execCtx.Session != nil {
-		parentSessionID = strings.TrimSpace(execCtx.Session.ID)
-	}
-
-	childRunner := r.newSubAgentChildRunner(workspace, preflight.Definition.MaxTurns)
-	if childRunner == nil || childRunner.client == nil {
-		return subAgentFailureResult(
-			invocationID,
-			agent,
-			subAgentErrorCodeRuntimeUnavailable,
-			"llm client is unavailable for subagent execution",
-			true,
-		)
-	}
-	childSession := newSubAgentSession(workspace, parentSessionID, invocationID, runMode)
-	defer childRunner.clearSessionSkillBridges(childSession)
-
-	if err := childRunner.syncExtensionTools(ctx, false); err != nil {
-		if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
-			return subAgentFailureResult(invocationID, agent, subAgentErrorCodeExecutionFailed, err.Error(), true)
-		}
-	}
-
-	userInput := buildSubAgentTaskInput(request)
-	setup, err := childRunner.prepareRunPrompt(childSession, RunPromptInput{
-		UserMessage: llm.NewUserTextMessage(userInput),
-		DisplayText: userInput,
-		SubAgent:    buildSubAgentPromptInput(request, preflight),
-	}, string(runMode))
-	if err != nil {
-		return subAgentFailureResult(invocationID, agent, subAgentErrorCodeExecutionFailed, err.Error(), true)
-	}
-	applySubAgentPreflightSetup(&setup, preflight)
-
-	answer, runErr := (&defaultEngine{runner: childRunner}).runPromptTurns(ctx, childSession, setup, nil)
-	if runErr != nil {
-		return subAgentFailureResult(invocationID, agent, subAgentErrorCodeExecutionFailed, runErr.Error(), true)
-	}
-
-	return buildSubAgentResultFromAnswer(answer, invocationID, agent)
-}
-
-func buildSubAgentResultFromAnswer(answer, invocationID, agent string) tools.DelegateSubAgentResult {
-	trimmed := strings.TrimSpace(answer)
-	if trimmed == "" {
-		trimmed = "SubAgent task completed."
-	}
-	base := tools.DelegateSubAgentResult{
-		OK:           true,
-		Status:       subAgentResultStatusCompleted,
-		InvocationID: strings.TrimSpace(invocationID),
-		Agent:        strings.TrimSpace(agent),
-	}
-
-	// Try to extract JSON from the answer. The LLM may wrap it in markdown fences
-	// or include prose before/after the JSON object.
-	jsonStr := extractJSONFromAnswer(trimmed)
-	if jsonStr == "" {
-		base.Summary = trimmed
-		return base
-	}
-
-	var parsed tools.DelegateSubAgentResult
-	if err := json.Unmarshal([]byte(jsonStr), &parsed); err != nil {
-		base.Summary = trimmed
-		return base
-	}
-
-	// Use parsed structured data if it contains meaningful content.
-	hasStructuredData := strings.TrimSpace(parsed.Summary) != ""
-	if !hasStructuredData {
-		base.Summary = trimmed
-		return base
-	}
-
-	base.Summary = strings.TrimSpace(parsed.Summary)
-	if base.Summary == "" {
-		base.Summary = trimmed
-	}
-	return base
-}
-
-// extractJSONFromAnswer attempts to find a JSON object in the answer text.
-// It handles markdown code fences and prose surrounding the JSON.
-func extractJSONFromAnswer(answer string) string {
-	// Try direct parse first (answer is pure JSON).
-	if json.Valid([]byte(answer)) {
-		return answer
-	}
-
-	// Try to extract from markdown code fences.
-	if idx := strings.Index(answer, "```"); idx >= 0 {
-		fenceEnd := strings.Index(answer[idx+3:], "\n")
-		if fenceEnd >= 0 {
-			codeStart := idx + 3 + fenceEnd + 1
-			codeEnd := strings.Index(answer[codeStart:], "```")
-			if codeEnd >= 0 {
-				candidate := strings.TrimSpace(answer[codeStart : codeStart+codeEnd])
-				if json.Valid([]byte(candidate)) {
-					return candidate
-				}
-			}
-		}
-	}
-
-	// Try to find the outermost { ... } pair.
-	first := strings.Index(answer, "{")
-	if first < 0 {
-		return ""
-	}
-	depth := 0
-	for i := first; i < len(answer); i++ {
-		switch answer[i] {
-		case '{':
-			depth++
-		case '}':
-			depth--
-			if depth == 0 {
-				candidate := answer[first : i+1]
-				if json.Valid([]byte(candidate)) {
-					return candidate
-				}
-			}
-		}
-	}
-	return ""
-}
-
-func (r *Runner) newSubAgentChildRunner(workspace string, maxTurns int) *Runner {
-	cfg := r.config
-	cfg.MaxIterations = resolveSubAgentMaxIterations(cfg.MaxIterations, maxTurns)
-	return NewRunner(Options{
-		Workspace:       workspace,
-		Config:          cfg,
-		Client:          r.GetClient(),
-		Registry:        r.registry,
-		Executor:        r.executor,
-		PolicyGateway:   r.policyGateway,
-		TaskManager:     r.taskManager,
-		Runtime:         r.runtime,
-		Extensions:      r.extensions,
-		SubAgentManager: r.subAgentManager,
-		TokenManager:    r.tokenManager,
-		AuditStore:      r.auditStore,
-		PromptStore:     r.promptStore,
-		Observer:        &noOpObserver{},
-		Approval:        nonInteractiveApproval(),
-		Stdin:           nil,
-		Stdout:          subAgentStdout(),
-	})
-}
-
-func newSubAgentSession(workspace, parentSessionID, invocationID string, runMode planpkg.AgentMode) *session.Session {
-	child := session.New(workspace)
-	base := strings.TrimSpace(parentSessionID)
-	if base == "" {
-		base = "session"
-	}
-	child.ID = fmt.Sprintf("%s/subagent/%s", base, strings.TrimSpace(invocationID))
-	child.Mode = runMode
-	child.ActiveSkill = nil
-	return child
-}
-
-func buildSubAgentTaskInput(request tools.DelegateSubAgentRequest) string {
-	task := strings.TrimSpace(request.Task)
-	if task == "" {
-		task = "Complete the delegated subagent task."
-	}
-	return task
-}
-
-func buildSubAgentPromptInput(request tools.DelegateSubAgentRequest, preflight subagentspkg.PreflightResult) *SubAgentPromptInput {
-	return &SubAgentPromptInput{
-		Name:           strings.TrimSpace(preflight.Definition.Name),
-		Task:           strings.TrimSpace(request.Task),
-		ScopePaths:     normalizeUniqueStrings(request.Scope.Paths),
-		ScopeSymbols:   normalizeUniqueStrings(request.Scope.Symbols),
-		AllowedTools:   append([]string(nil), preflight.EffectiveTools...),
-		Isolation:      strings.TrimSpace(preflight.Isolation),
-		ResultPolicy:   subAgentResultPolicyCompressed,
-		DefinitionBody: strings.TrimSpace(preflight.Definition.Instruction),
-	}
-}
-
-func applySubAgentPreflightSetup(setup *runPromptSetup, preflight subagentspkg.PreflightResult) {
-	if setup == nil {
-		return
-	}
-	setup.AllowedTools = cloneToolSet(preflight.AllowedTools)
-	setup.DeniedTools = cloneToolSet(preflight.DeniedTools)
-	setup.AllowedToolNames = append([]string(nil), preflight.EffectiveTools...)
-	setup.DeniedToolNames = sortedToolSetNames(preflight.DeniedTools)
-	setup.AvailableTools = append([]string(nil), preflight.EffectiveTools...)
-	setup.AvailableSubAgents = nil
-	setup.ActiveSkill = nil
-}
-
-func sortedToolSetNames(set map[string]struct{}) []string {
-	if len(set) == 0 {
-		return nil
-	}
-	names := make([]string, 0, len(set))
-	for name := range set {
-		trimmed := strings.TrimSpace(name)
-		if trimmed == "" {
-			continue
-		}
-		names = append(names, trimmed)
-	}
-	if len(names) == 0 {
-		return nil
-	}
-	slices.Sort(names)
-	return names
-}
-
-func resolveSubAgentMaxIterations(parentMaxIterations int, requestedMaxTurns int) int {
-	effectiveParent := parentMaxIterations
-	if effectiveParent <= 0 {
-		effectiveParent = defaultSubAgentMaxIterations
-	}
-	if requestedMaxTurns > 0 && requestedMaxTurns < effectiveParent {
-		return requestedMaxTurns
-	}
-	return effectiveParent
-}
-
-func subAgentFailureResult(invocationID, agent, code, message string, retryable bool) tools.DelegateSubAgentResult {
-	return tools.DelegateSubAgentResult{
-		OK:           false,
-		Status:       subAgentResultStatusFailed,
-		InvocationID: strings.TrimSpace(invocationID),
-		Agent:        strings.TrimSpace(agent),
-		Error: &tools.DelegateSubAgentError{
-			Code:      strings.TrimSpace(code),
-			Message:   strings.TrimSpace(message),
-			Retryable: retryable,
-		},
-	}
 }
 
 func preflightResultName(agent string) string {
